+++ react/components/ProductBrand/ProductBrandName.tsx
@@ -10,7 +10,7 @@ interface Props {
 
 export const CSS_HANDLES = ['productBrandNameLink', 'productBrandName'] as const
 
-function ProductBrandName({ brandName, withLink, slug }: Props) {
+function ProductBrandName({ brandName, withLink, slug}: Props) {
   const { handles } = useProductBrandCssHandles()
 
   if (withLink && slug) {
@@ -22,7 +22,7 @@ function ProductBrandName({ brandName, withLink, slug }: Props) {
         className={`${handles.productBrandNameLink}`}
         data-testid="name-redirect"
       >
-        <span className={`${handles.productBrandName}`}>{brandName}</span>
+        <span className={`${handles.productBrandName}`} aria-label={"Brand " + brandName}>{brandName}</span>
       </a>
     )
   }